Updated component and navigation_component directories. Moved shapes directory.
