+++ app/src/main/java/com/koleff/kare_android/ui/compose/scaffolds/ExerciseDetailsConfiguratorScreenScaffold.kt
@@ -11,13 +11,11 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.MutableState
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.unit.dp
-import androidx.navigation.NavHostController
-import com.koleff.kare_android.ui.compose.navigation.shapes.RoundedToolbarShape
+import com.koleff.kare_android.ui.compose.shapes.RoundedToolbarShape
 
 @Composable
 fun ExerciseDetailsConfiguratorScreenScaffold(
