+++ app/src/main/java/com/koleff/kare_android/ui/compose/scaffolds/SearchListScaffold.kt
@@ -4,11 +4,8 @@ import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.MutableState
 import androidx.compose.ui.Modifier
-import androidx.navigation.NavHostController
-import com.koleff.kare_android.ui.compose.navigation.BottomNavigationBar
-import com.koleff.kare_android.ui.compose.navigation.Toolbar
+import com.koleff.kare_android.ui.compose.components.navigation_components.Toolbar
 
 @Composable
 fun SearchListScaffold(
