+++ app/src/main/java/com/koleff/kare_android/ui/compose/screen/ExerciseDetailsScreen.kt
@@ -33,8 +33,8 @@ import com.koleff.kare_android.common.navigation.NavigationEvent
 import com.koleff.kare_android.data.model.dto.ExerciseDetailsDto
 import com.koleff.kare_android.data.model.dto.MachineType
 import com.koleff.kare_android.data.model.dto.MuscleGroup
-import com.koleff.kare_android.ui.compose.LoadingWheel
-import com.koleff.kare_android.ui.compose.YoutubeVideoPlayer
+import com.koleff.kare_android.ui.compose.components.LoadingWheel
+import com.koleff.kare_android.ui.compose.components.YoutubeVideoPlayer
 import com.koleff.kare_android.ui.compose.scaffolds.ExerciseDetailsScreenScaffold
 import com.koleff.kare_android.ui.state.ExerciseDetailsState
 import com.koleff.kare_android.ui.view_model.ExerciseDetailsViewModel
