+++ app/src/main/java/com/koleff/kare_android/ui/compose/screen/SearchWorkoutsScreen.kt
@@ -22,9 +22,9 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.koleff.kare_android.common.navigation.Destination
 import com.koleff.kare_android.common.navigation.NavigationEvent
-import com.koleff.kare_android.ui.compose.LoadingWheel
-import com.koleff.kare_android.ui.compose.SearchBar
-import com.koleff.kare_android.ui.compose.SearchWorkoutList
+import com.koleff.kare_android.ui.compose.components.LoadingWheel
+import com.koleff.kare_android.ui.compose.components.SearchBar
+import com.koleff.kare_android.ui.compose.components.SearchWorkoutList
 import com.koleff.kare_android.ui.compose.scaffolds.SearchListScaffold
 import com.koleff.kare_android.ui.view_model.ExerciseViewModel
 import com.koleff.kare_android.ui.view_model.SearchWorkoutViewModel
